feat: comprehensive test suite improvement and BangumiParser implementation
## New Features
- BangumiParser: Robust anime title parsing with TDD approach
  - Resolution extraction (1080p, 720p, 4K, etc.)
  - Codec detection (HEVC, AVC, FLAC, AAC, etc.)
  - Subtitle language detection with SubtitleType enum
  - Season extraction with Chinese numeral support (第一季, S1, etc.)
  - Episode extraction with batch range support (01-12)
  - Source/rip type detection (BDRip, WEBRip, etc.)
  - Title extraction with multilingual support
  - Movie/OVA/Special detection (★剧场版, 劇場版, etc.)

## Test Suite Improvements
- Created shared test fixtures in conftest.py
- Added pytest markers configuration (slow, integration, unit)
- Parametrized all extraction test classes for better coverage
- Added edge case tests for titles, episodes, seasons, brackets
- Created RSS parsing integration tests with real feed fixtures
- Added performance tests and optimization benchmarks
- Improved test consistency across test_raw_parser.py and integration tests

## Bug Fixes
- Fixed 7 parser bugs discovered during test suite development
- Added support for movie release decorator parsing

## Files Added
- backend/src/module/models/parsed.py (ParsedBangumi model)
- backend/src/module/parser/analyser/bangumi_parser.py (1298 lines)
- backend/src/test/conftest.py (shared fixtures)
- backend/src/test/test_bangumi_parser.py (3868 lines, 200+ test cases)
- backend/src/test/test_parsed_model.py
- backend/src/test/test_rss_parsing_integration.py
- backend/src/test/test_torrent_parser.py
- Test fixtures: aggregate_rss.xml, edge_case_titles.json

+++ backend/src/module/models/torrent.py
@@ -24,7 +24,7 @@ class EpisodeFile(BaseModel):
     group: str | None = Field(None)
     title: str = Field(...)
     season: int = Field(...)
-    episode: int | float = Field(None)
+    episode: float | int = Field(None)
     suffix: str = Field(..., regex=r"\.(mkv|mp4|MKV|MP4)$")
 
 
@@ -33,6 +33,6 @@ class SubtitleFile(BaseModel):
     group: str | None = Field(None)
     title: str = Field(...)
     season: int = Field(...)
-    episode: int | float = Field(None)
+    episode: float | int = Field(None)
     language: str = Field(..., regex=r"(zh|zh-tw)")
     suffix: str = Field(..., regex=r"\.(ass|srt|ASS|SRT)$")
